+++ src/main/java/lonestarrr/arconia/common/core/proxy/ClientProxy.java
@@ -1,17 +1,13 @@
 package lonestarrr.arconia.common.core.proxy;
 
-import lonestarrr.arconia.client.core.handler.ColorHandler;
 import lonestarrr.arconia.client.core.handler.BlockEntityRendererHandler;
+import lonestarrr.arconia.client.core.handler.ColorHandler;
 import lonestarrr.arconia.client.effects.BuildPatternPreview;
 import lonestarrr.arconia.client.effects.PotItemTransfers;
 import lonestarrr.arconia.client.gui.render.HighlightPatternStructure;
 import lonestarrr.arconia.common.Arconia;
-import lonestarrr.arconia.common.block.ModBlocks;
-import lonestarrr.arconia.common.core.RainbowColor;
 import lonestarrr.arconia.common.item.MagicInABottle;
 import lonestarrr.arconia.common.item.ModItems;
-import net.minecraft.client.renderer.ItemBlockRenderTypes;
-import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.client.event.RegisterColorHandlersEvent;
@@ -44,20 +40,6 @@ public class ClientProxy implements IProxy {
         Arconia.logger.info("********************* client-side proxy init");
 
         event.enqueueWork(() -> {
-            // TODO Can these blocks themselves provide this hint?
-            RenderType cutout = RenderType.cutout();
-
-            // TODO See McJTY porting guide on how to set render_type in the models
-            // https://www.mcjty.eu/docs/1.19/porting
-            ItemBlockRenderTypes.setRenderLayer(ModBlocks.clover.get(), cutout);
-
-            for (RainbowColor tier : RainbowColor.values()) {
-                ItemBlockRenderTypes.setRenderLayer(ModBlocks.getArconiumTreeLeaves(tier).get(), cutout);
-                ItemBlockRenderTypes.setRenderLayer(ModBlocks.getArconiumTreeSapling(tier).get(), cutout);
-                // gleaned from Blocks.GRASS_BLOCK - this is for overlaying the top with a rainbow tint
-                ItemBlockRenderTypes.setRenderLayer(ModBlocks.getInfiniteGoldArconiumBlock(tier).get(), RenderType.cutoutMipped());
-            }
-
             registerItemProperties();
         });
     }